LayoutTests/imported/w3c/web-platform-tests/editing/other/merge-span-with-style-after-pressing-enter-followed-by-backspace-in-contenteditable-div-expected.txt

Back

pace


FAIL waiting for command to execute assert_equals: Style is not preserved for the span after pressing backspace in contenteditable expected "<h1><span style=\"background-color:red;\">Back</span><span style=\"background-color: red;\">space</span></h1>" but got "<h1><span style=\"background-color:red;\">Back</span></h1><h1><span style=\"background-color:red;\">pace</span></h1>"
